@@ -18,18 +18,17 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  cached_network_image: ^2.3.3
-  webview_flutter: ^1.0.7
+  http: ^0.13.0
+  cached_network_image: ^3.0.0
+  webview_flutter: ^2.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-
+  cupertino_icons: ^1.0.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
