@@ -6,18 +6,19 @@ import 'article_view.dart';
 
 class CategoryNews extends StatefulWidget {
   final String category;
-  CategoryNews({this.category});
+
+  const CategoryNews({required this.category});
+
   @override
   _CategoryNewsState createState() => _CategoryNewsState();
 }
 
 class _CategoryNewsState extends State<CategoryNews> {
-
-  List<ArticleModel> articles = new List<ArticleModel>();
+  List<ArticleModel> articles = [];
   bool _loading = true;
+
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
     getCategoryNews();
   }
@@ -30,102 +31,117 @@ class _CategoryNewsState extends State<CategoryNews> {
       _loading = false;
     });
   }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Row(
           mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
+          children: const <Widget>[
             Text("My"),
             Text("Crypto"),
             Text(
               "News",
               style: TextStyle(color: Colors.blue),
-            )
+            ),
           ],
         ),
         actions: <Widget>[
           Opacity(
-              opacity: 0,
-              child: Container(
-                padding: EdgeInsets.symmetric(horizontal: 16),
-                child: Icon(Icons.save),
-              )
-
+            opacity: 0,
+            child: Container(
+              padding: EdgeInsets.symmetric(horizontal: 16),
+              child: Icon(Icons.save),
+            ),
           )
         ],
         centerTitle: true,
         elevation: 0.0,
       ),
-      body:  _loading
+      body: _loading
           ? Center(
-        child: Container(
-          child: CircularProgressIndicator(),
-        ),
-      )
-          :  SingleChildScrollView(
-            child: Container(
-              padding: EdgeInsets.symmetric(horizontal: 16),
-        child: Column(
-            children: <Widget>[
-              ///Blog
-              Container(
-                padding: EdgeInsets.only(top: 16),
-                child: ListView.builder(
-                    itemCount: articles.length,
-                    shrinkWrap: true,
-                    physics: ClampingScrollPhysics(),
-                    itemBuilder: (context, index) {
-                      return BlogTile(
-                        imageUrl: articles[index].urlToImage,
-                        title: articles[index].title,
-                        desc: articles[index].description,
-                        url: articles[index].url,
-                      );
-                    }),
-              )
-            ],
-        ),
-      ),
-          ),
+              child: Container(
+                child: const CircularProgressIndicator(),
+              ),
+            )
+          : SingleChildScrollView(
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Column(
+                  children: <Widget>[
+                    ///Blog
+                    Container(
+                      padding: const EdgeInsets.only(top: 16),
+                      child: ListView.builder(
+                        itemCount: articles.length,
+                        shrinkWrap: true,
+                        physics: const ClampingScrollPhysics(),
+                        itemBuilder: (context, index) {
+                          return BlogTile(
+                            imageUrl: articles[index].urlToImage,
+                            title: articles[index].title,
+                            desc: articles[index].description,
+                            url: articles[index].url,
+                          );
+                        },
+                      ),
+                    )
+                  ],
+                ),
+              ),
+            ),
     );
   }
 }
 
 class BlogTile extends StatelessWidget {
   final String imageUrl, title, desc, url;
-  BlogTile(
-      {@required this.imageUrl, @required this.title, @required this.desc, @required this.url});
+
+  const BlogTile({
+    required this.imageUrl,
+    required this.title,
+    required this.desc,
+    required this.url,
+  });
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: (){
-        Navigator.push(context, MaterialPageRoute(
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
             builder: (context) => ArticleView(
               blogUrl: url,
-            )
-        ));
+            ),
+          ),
+        );
       },
       child: Container(
-        margin: EdgeInsets.only(bottom: 16),
+        margin: const EdgeInsets.only(bottom: 16),
         child: Column(
           children: <Widget>[
             ClipRRect(
-                borderRadius: BorderRadius.circular(6),
-                child: Image.network(imageUrl)),
-            SizedBox(height: 8,),
-            Text(title, style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.w500,
-              color: Colors.black87,
-            ),),
-            SizedBox(height: 8,),
-            Text(desc, style: TextStyle(
-              color: Colors.black54,
-            ),)
-
+              borderRadius: BorderRadius.circular(6),
+              child: Image.network(imageUrl),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              title,
+              style: const TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.w500,
+                color: Colors.black87,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              desc,
+              style: const TextStyle(
+                color: Colors.black54,
+              ),
+            ),
           ],
         ),
       ),
