@@ -1,5 +1,5 @@
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/cupertino.dart';
+// import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:mycryptonews/helper/data.dart';
 import 'package:mycryptonews/helper/news.dart';
@@ -15,14 +15,13 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-  List<CategoryModel> categories = new List<CategoryModel>();
-  List<ArticleModel> articles = new List<ArticleModel>();
+  List<CategoryModel> categories = [];
+  List<ArticleModel> articles = [];
 
   bool _loading = true;
 
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
     categories = getCategories();
     getNews();
@@ -43,7 +42,7 @@ class _HomeState extends State<Home> {
       appBar: AppBar(
         title: Row(
           mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
+          children: const <Widget>[
             Text("My"),
             Text("Crypto"),
             Text(
@@ -57,44 +56,46 @@ class _HomeState extends State<Home> {
       body: _loading
           ? Center(
               child: Container(
-                child: CircularProgressIndicator(),
+                child: const CircularProgressIndicator(),
               ),
             )
           : SingleChildScrollView(
               child: Container(
-                padding: EdgeInsets.symmetric(horizontal: 16),
+                padding: const EdgeInsets.symmetric(horizontal: 16),
                 child: Column(
                   children: <Widget>[
                     ///Categories
-                    Container(
+                    SizedBox(
                       height: 70,
                       child: ListView.builder(
-                          itemCount: categories.length,
-                          shrinkWrap: true,
-                          scrollDirection: Axis.horizontal,
-                          itemBuilder: (context, index) {
-                            return CategoryTile(
-                              imageUrl: categories[index].imageUrl,
-                              categoryName: categories[index].categoryName,
-                            );
-                          }),
+                        itemCount: categories.length,
+                        shrinkWrap: true,
+                        scrollDirection: Axis.horizontal,
+                        itemBuilder: (context, index) {
+                          return CategoryTile(
+                            imageUrl: categories[index].imageUrl,
+                            categoryName: categories[index].categoryName,
+                          );
+                        },
+                      ),
                     ),
 
                     ///Blog
                     Container(
-                      padding: EdgeInsets.only(top: 16),
+                      padding: const EdgeInsets.only(top: 16),
                       child: ListView.builder(
-                          itemCount: articles.length,
-                          shrinkWrap: true,
-                          physics: ClampingScrollPhysics(),
-                          itemBuilder: (context, index) {
-                            return BlogTile(
-                                imageUrl: articles[index].urlToImage,
-                                title: articles[index].title,
-                                desc: articles[index].description,
-                              url: articles[index].url,
-                            );
-                          }),
+                        itemCount: articles.length,
+                        shrinkWrap: true,
+                        physics: const ClampingScrollPhysics(),
+                        itemBuilder: (context, index) {
+                          return BlogTile(
+                            imageUrl: articles[index].urlToImage,
+                            title: articles[index].title,
+                            desc: articles[index].description,
+                            url: articles[index].url,
+                          );
+                        },
+                      ),
                     )
                   ],
                 ),
@@ -105,30 +106,36 @@ class _HomeState extends State<Home> {
 }
 
 class CategoryTile extends StatelessWidget {
-  final imageUrl, categoryName;
-  CategoryTile({this.imageUrl, this.categoryName});
+  final String imageUrl, categoryName;
+
+  const CategoryTile({required this.imageUrl, required this.categoryName});
+
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () {
-        Navigator.push(context, MaterialPageRoute(
+        Navigator.push(
+          context,
+          MaterialPageRoute(
             builder: (context) => CategoryNews(
               category: categoryName.toLowerCase(),
-            )
-        ));
+            ),
+          ),
+        );
       },
       child: Container(
-        margin: EdgeInsets.only(right: 16),
+        margin: const EdgeInsets.only(right: 16),
         child: Stack(
           children: <Widget>[
             ClipRRect(
-                borderRadius: BorderRadius.circular(6),
-                child: CachedNetworkImage(
-                  imageUrl: imageUrl,
-                  width: 120,
-                  height: 80,
-                  fit: BoxFit.cover,
-                )),
+              borderRadius: BorderRadius.circular(6),
+              child: CachedNetworkImage(
+                imageUrl: imageUrl,
+                width: 120,
+                height: 80,
+                fit: BoxFit.cover,
+              ),
+            ),
             Container(
               alignment: Alignment.center,
               width: 120,
@@ -139,12 +146,13 @@ class CategoryTile extends StatelessWidget {
               ),
               child: Text(
                 categoryName,
-                style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 14,
-                    fontWeight: FontWeight.w500),
+                style: const TextStyle(
+                  color: Colors.white,
+                  fontSize: 14,
+                  fontWeight: FontWeight.w500,
+                ),
               ),
-            )
+            ),
           ],
         ),
       ),
@@ -154,37 +162,51 @@ class CategoryTile extends StatelessWidget {
 
 class BlogTile extends StatelessWidget {
   final String imageUrl, title, desc, url;
-  BlogTile(
-      {@required this.imageUrl, @required this.title, @required this.desc, @required this.url});
+
+  const BlogTile({
+    required this.imageUrl,
+    required this.title,
+    required this.desc,
+    required this.url,
+  });
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: (){
-        Navigator.push(context, MaterialPageRoute(
+      onTap: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
             builder: (context) => ArticleView(
               blogUrl: url,
-            )
-        ));
+            ),
+          ),
+        );
       },
       child: Container(
-        margin: EdgeInsets.only(bottom: 16),
+        margin: const EdgeInsets.only(bottom: 16),
         child: Column(
           children: <Widget>[
             ClipRRect(
-                borderRadius: BorderRadius.circular(6),
-                child: Image.network(imageUrl)),
-            SizedBox(height: 8,),
-            Text(title, style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.w500,
-              color: Colors.black87,
-          ),),
-            SizedBox(height: 8,),
-            Text(desc, style: TextStyle(
-              color: Colors.black54,
-            ),)
-
+              borderRadius: BorderRadius.circular(6),
+              child: Image.network(imageUrl),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              title,
+              style: const TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.w500,
+                color: Colors.black87,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              desc,
+              style: const TextStyle(
+                color: Colors.black54,
+              ),
+            ),
           ],
         ),
       ),
