@@ -1,4 +1,4 @@
-package online.thephiri.mycryptonews
+package online.thephiri.my_crypto_news
 
 import io.flutter.embedding.android.FlutterActivity
 
