@@ -1,76 +1,84 @@
 import 'package:mycryptonews/models/category_model.dart';
 
-List<CategoryModel> getCategories(){
-  List<CategoryModel> category = new List<CategoryModel>();
+List<CategoryModel> getCategories() {
+  List<CategoryModel> categories = [];
 
+  CategoryModel categoryModel = CategoryModel(
+    categoryName: "Bitcoin (BTC)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //1
-  CategoryModel categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Bitcoin (BTC)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Ethereum (ETH)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //2
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Ethereum (ETH)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Ripple (XRP)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //3
-   categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Ripple (XRP)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Litecoin (LTC)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //4
-   categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Litecoin (LTC)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Tether (USDT)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //5
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Tether (USDT)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Bitcoin Cash (BCH)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //6
-   categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Bitcoin Cash (BCH)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Libra (LIBRA)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //7
-   categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Libra (LIBRA)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Monero (XMR)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //8
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Monero (XMR)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "EOS (EOS)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //9
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "EOS (EOS)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Bitcoin SV (BSV)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //10
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Bitcoin SV (BSV)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
+  categoryModel = CategoryModel(
+    categoryName: "Binance Coin (BNB)",
+    imageUrl:
+        "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000",
+  );
+  categories.add(categoryModel);
 
-  //11
-  categoryModel = new CategoryModel();
-  categoryModel.categoryName = "Binance Coin (BNB)";
-  categoryModel.imageUrl = "https://images.unsplash.com/photo-1554260570-e9689a3418b8?ixlib=rb-1.2.1&ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&auto=format&fit=crop&w=1360&q=8000";
-  category.add(categoryModel);
-
-  return category;
-
-  
+  return categories;
 }
