@@ -1,4 +1,6 @@
-class CategoryModel{
+class CategoryModel {
   String categoryName;
   String imageUrl;
+
+  CategoryModel({required this.categoryName, required this.imageUrl});
 }
