@@ -5,14 +5,16 @@ import 'package:webview_flutter/webview_flutter.dart';
 
 class ArticleView extends StatefulWidget {
   final String blogUrl;
-  ArticleView({this.blogUrl});
+
+  const ArticleView({required this.blogUrl});
+
   @override
   _ArticleViewState createState() => _ArticleViewState();
 }
 
 class _ArticleViewState extends State<ArticleView> {
-
-  final Completer<WebViewController> _completer = Completer<WebViewController>();
+  final Completer<WebViewController> _completer =
+      Completer<WebViewController>();
 
   @override
   Widget build(BuildContext context) {
@@ -20,36 +22,35 @@ class _ArticleViewState extends State<ArticleView> {
       appBar: AppBar(
         title: Row(
           mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
+          children: const <Widget>[
             Text("My"),
             Text("Crypto"),
             Text(
               "News",
               style: TextStyle(color: Colors.blue),
-            )
+            ),
           ],
         ),
         actions: <Widget>[
-         Opacity(
-        opacity: 0,
-        child: Container(
-          padding: EdgeInsets.symmetric(horizontal: 16),
-          child: Icon(Icons.save),
-      )
-
-        )
+          Opacity(
+            opacity: 0,
+            child: Container(
+              padding: EdgeInsets.symmetric(horizontal: 16),
+              child: Icon(Icons.save),
+            ),
+          )
         ],
         centerTitle: true,
         elevation: 0.0,
       ),
       body: Container(
         height: MediaQuery.of(context).size.height,
-        width: MediaQuery.of(context).size.height,
+        width: MediaQuery.of(context).size.width,
         child: WebView(
           initialUrl: widget.blogUrl,
-          onWebViewCreated: ((WebViewController webViewController){
+          onWebViewCreated: (WebViewController webViewController) {
             _completer.complete(webViewController);
-          }),
+          },
         ),
       ),
     );
